Started adding url functionality and reorganized files
