+++ site/client/owid.entry.ts
@@ -29,7 +29,7 @@ import { runLightbox } from "./Lightbox"
 import { runSiteTools } from "./SiteTools"
 import { runCovid } from "./covid/index"
 import { runGlobalEntityControl } from "./global-entity/GlobalEntityControl"
-import { CovidChartBuilder } from "charts/CovidChartBuilder"
+import { CovidChartBuilder } from "charts/covidChartBuilder/CovidChartBuilder"
 
 declare var window: any
 window.Grapher = Grapher
